requirements.txt
# Django Framework
Django>=5.2,<5.3
djangorestframework>=3.15.0,<4.0.0

# CORS Headers for React integration
django-cors-headers>=4.4.0,<5.0.0

# Environment variable management
python-decouple>=3.8,<4.0

# Database
# SQLite is included with Python, no additional package needed

# Optional but recommended for production
gunicorn>=21.2.0,<22.0.0  # WSGI HTTP Server
whitenoise>=6.6.0,<7.0.0  # Static file serving
